+++ game/public/data/manifests/manifest-schema.json
@@ -1,0 +1,240 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "title": "Otterblade Asset Manifest Schema",
+  "description": "Schema for asset generation manifests with approval workflow",
+  "type": "object",
+  "required": ["category", "provider", "outputDir", "assets"],
+  "properties": {
+    "$schema": {
+      "type": "string"
+    },
+    "category": {
+      "type": "string",
+      "enum": ["sprites", "enemies", "cinematics", "scenes", "sounds", "chapter-plates", "items", "effects"],
+      "description": "Asset category for grouping"
+    },
+    "provider": {
+      "type": "string",
+      "enum": ["openai", "google", "freesound", "custom"],
+      "description": "Generation provider"
+    },
+    "model": {
+      "type": "string",
+      "description": "Model identifier for the provider"
+    },
+    "outputDir": {
+      "type": "string",
+      "description": "Target directory for generated assets (relative to repo root)"
+    },
+    "notes": {
+      "type": "string",
+      "description": "Human-readable notes about this manifest"
+    },
+    "assets": {
+      "type": "array",
+      "items": {
+        "$ref": "#/definitions/asset"
+      }
+    }
+  },
+  "definitions": {
+    "asset": {
+      "type": "object",
+      "required": ["id", "name", "filename", "status"],
+      "properties": {
+        "id": {
+          "type": "string",
+          "pattern": "^[a-z0-9_]+$",
+          "description": "Unique identifier for this asset"
+        },
+        "name": {
+          "type": "string",
+          "description": "Human-readable name"
+        },
+        "filename": {
+          "type": "string",
+          "description": "Output filename"
+        },
+        "status": {
+          "type": "string",
+          "enum": ["pending", "generating", "complete", "needs_regeneration", "approved", "rejected"],
+          "description": "Current workflow status"
+        },
+        "reason": {
+          "type": "string",
+          "description": "Reason for current status (especially for needs_regeneration or rejected)"
+        },
+        "type": {
+          "type": "string",
+          "description": "Asset subtype within category"
+        },
+        "config": {
+          "type": "object",
+          "description": "Provider-specific configuration",
+          "properties": {
+            "duration": {
+              "type": ["number", "string"],
+              "description": "Duration for video/audio assets"
+            },
+            "aspectRatio": {
+              "type": "string",
+              "description": "Aspect ratio for images/videos"
+            },
+            "includeAudio": {
+              "type": "boolean",
+              "description": "Whether to include audio in video generation"
+            },
+            "stereo": {
+              "type": "boolean",
+              "description": "Stereo audio for sound assets"
+            },
+            "width": {
+              "type": "number"
+            },
+            "height": {
+              "type": "number"
+            },
+            "frameCount": {
+              "type": "number",
+              "description": "Number of frames for sprite sheets"
+            },
+            "frameRate": {
+              "type": "number",
+              "description": "Frames per second for animations"
+            }
+          }
+        },
+        "prompt": {
+          "type": "object",
+          "description": "Generation prompt details",
+          "additionalProperties": true
+        },
+        "validation": {
+          "type": "object",
+<<<<<<< HEAD
+          "description": "Validation rules for brand compliance and file requirements",
+          "properties": {
+            "noHumans": { "type": "boolean" },
+            "protagonistIsOtter": { "type": "boolean" },
+            "warmPalette": { "type": "boolean" },
+            "minFileSize": {
+              "type": "integer",
+              "minimum": 0,
+              "description": "Minimum file size in bytes"
+            },
+            "maxFileSize": {
+              "type": "integer",
+              "minimum": 0,
+              "description": "Maximum file size in bytes"
+            }
+=======
+          "description": "Validation rules for brand compliance",
+          "properties": {
+            "noHumans": { "type": "boolean" },
+            "protagonistIsOtter": { "type": "boolean" },
+            "warmPalette": { "type": "boolean" }
+>>>>>>> main
+          }
+        },
+        "variants": {
+          "type": "array",
+          "description": "Alternative prompts for regeneration attempts",
+          "items": {
+            "type": "object",
+            "properties": {
+              "id": {
+                "type": "string",
+                "description": "Variant identifier (e.g., 'v2', 'dramatic', 'warm')"
+              },
+              "prompt": {
+                "type": "object",
+                "description": "Override prompt for this variant"
+              },
+              "generatedAt": {
+                "type": "string",
+                "format": "date-time"
+              },
+              "approved": {
+                "type": "boolean"
+              }
+            }
+          }
+        },
+        "generationHistory": {
+          "type": "array",
+          "description": "History of generation attempts",
+          "items": {
+            "type": "object",
+            "properties": {
+              "timestamp": {
+                "type": "string",
+                "format": "date-time"
+              },
+              "provider": { "type": "string" },
+              "model": { "type": "string" },
+              "promptHash": {
+                "type": "string",
+                "description": "SHA256 hash of prompt for deduplication"
+              },
+              "success": { "type": "boolean" },
+              "error": { "type": "string" },
+              "variantId": { "type": "string" }
+            }
+          }
+        },
+        "approval": {
+          "type": "object",
+          "description": "Approval workflow state",
+          "properties": {
+            "approved": {
+              "type": "boolean",
+              "description": "Whether asset is approved for production"
+            },
+            "approvedAt": {
+              "type": "string",
+              "format": "date-time"
+            },
+            "approvedBy": {
+              "type": "string",
+              "description": "Identifier of who approved"
+            },
+            "notes": {
+              "type": "string",
+              "description": "Approval notes"
+            },
+            "variantUsed": {
+              "type": "string",
+              "description": "Which variant was approved (if any)"
+            }
+          }
+        },
+        "metadata": {
+          "type": "object",
+          "description": "Runtime metadata",
+          "properties": {
+            "fileSize": { "type": "number" },
+            "dimensions": {
+              "type": "object",
+              "properties": {
+                "width": { "type": "number" },
+                "height": { "type": "number" }
+              }
+            },
+            "duration": { "type": "number" },
+            "checksum": { "type": "string" }
+          }
+        },
+        "loop": {
+          "type": "boolean",
+          "description": "Whether audio should loop"
+        },
+        "volume": {
+          "type": "number",
+          "minimum": 0,
+          "maximum": 1,
+          "description": "Default volume for audio"
+        }
+      }
+    }
+  }
+}
