+++ game/public/data/manifests/scenes.json
@@ -1,0 +1,154 @@
+{
+  "$schema": "./manifest-schema.json",
+  "category": "scenes",
+  "provider": "google",
+  "model": "imagen-3.0-generate-002",
+  "outputDir": "game/src/assets/images",
+  "assets": [
+    {
+      "type": "scene",
+      "type": "scene",
+      "id": "parallax_village",
+      "name": "Village Morning Parallax",
+      "filename": "parallax/village_morning_parallax_background.png",
+      "status": "complete",
+      "config": {
+        "aspectRatio": "16:9",
+        "outputMimeType": "image/png"
+      },
+      "prompt": {
+        "scene": "Finn's home village at dawn",
+        "setting": "Small riverside hamlet with thatched cottages, willow trees trailing in the water, morning mist",
+        "composition": "Wide parallax format, clear depth layers (cottages foreground, river midground, forest background)",
+        "lighting": "Golden dawn light, soft mist, warm lanterns in windows",
+        "style": "Storybook painterly, warm and cozy, NO characters"
+      }
+    },
+    {
+      "type": "scene",
+      "id": "parallax_abbey_exterior",
+      "name": "Abbey Exterior Parallax",
+      "filename": "parallax/abbey_exterior_parallax_background.png",
+      "status": "complete",
+      "config": {
+        "aspectRatio": "16:9",
+        "outputMimeType": "image/png"
+      },
+      "prompt": {
+        "scene": "Willowmere Hearthhold exterior walls",
+        "setting": "Ancient stone walls with moss and ivy, gatehouse towers, forest approach",
+        "composition": "Parallax layers: trees foreground, wall midground, towers background",
+        "lighting": "Afternoon sun, god rays through canopy",
+        "style": "Storybook painterly, ancient and welcoming"
+      }
+    },
+    {
+      "type": "scene",
+      "id": "parallax_abbey_interior",
+      "name": "Abbey Interior Parallax",
+      "filename": "parallax/abbey_interior_parallax_background.png",
+      "status": "complete",
+      "config": {
+        "aspectRatio": "16:9",
+        "outputMimeType": "image/png"
+      },
+      "prompt": {
+        "scene": "Great Hall interior with the Everember",
+        "setting": "Vaulted stone hall, great hearth with eternal fire, tapestries, candlelight",
+        "composition": "Pillars foreground, Everember hearth midground, tapestries background",
+        "lighting": "Warm firelight, candle glow, amber and gold tones",
+        "style": "Storybook painterly, cozy and sacred"
+      }
+    },
+    {
+      "type": "scene",
+      "id": "parallax_dungeon",
+      "name": "Dungeon Parallax",
+      "filename": "parallax/dungeon_parallax_background.png",
+      "status": "complete",
+      "config": {
+        "aspectRatio": "16:9",
+        "outputMimeType": "image/png"
+      },
+      "prompt": {
+        "scene": "Deep Cellars underground passages",
+        "setting": "Stone corridors, torch sconces, roots breaking through ceiling, damp atmosphere",
+        "composition": "Archways foreground, passages midground, darkness background",
+        "lighting": "Torchlight pools, deep shadows, orange against blue-black",
+        "style": "Storybook painterly, tense but not horror"
+      }
+    },
+    {
+      "type": "scene",
+      "id": "parallax_courtyard",
+      "name": "Courtyard Parallax",
+      "filename": "parallax/courtyard_parallax_background.png",
+      "status": "complete",
+      "config": {
+        "aspectRatio": "16:9",
+        "outputMimeType": "image/png"
+      },
+      "prompt": {
+        "scene": "Kitchen Gardens and training yard",
+        "setting": "Sunlit courtyard, vegetable plots, herb spirals, training dummies, surrounding walls",
+        "composition": "Garden plots foreground, open yard midground, buildings background",
+        "lighting": "Bright daylight, clear blue sky, vibrant greens",
+        "style": "Storybook painterly, hopeful and alive"
+      }
+    },
+    {
+      "type": "scene",
+      "id": "parallax_rooftops",
+      "name": "Rooftops Parallax",
+      "filename": "parallax/rooftops_parallax_background.png",
+      "status": "complete",
+      "config": {
+        "aspectRatio": "16:9",
+        "outputMimeType": "image/png"
+      },
+      "prompt": {
+        "scene": "Bell Tower and rooftops at sunset",
+        "setting": "Slate rooftops, wind-worn banners, bell tower spire, storm clouds gathering",
+        "composition": "Chimneys foreground, rooflines midground, bell tower and sky background",
+        "lighting": "Dramatic sunset, orange and purple, wind-swept atmosphere",
+        "style": "Storybook painterly, dramatic and urgent"
+      }
+    },
+    {
+      "type": "scene",
+      "id": "parallax_outer_ruins",
+      "name": "Outer Ruins Parallax",
+      "filename": "parallax/outer_ruins_parallax_background.png",
+      "status": "complete",
+      "config": {
+        "aspectRatio": "16:9",
+        "outputMimeType": "image/png"
+      },
+      "prompt": {
+        "scene": "Storm-damaged outer ramparts",
+        "setting": "Crumbling battlements, fog rolling in, lightning in distance, frost on stones",
+        "composition": "Broken walls foreground, rampart walk midground, storm sky background",
+        "lighting": "Storm light, cold blues and grays, occasional lightning flash",
+        "style": "Storybook painterly, foreboding but not horror"
+      }
+    },
+    {
+      "type": "scene",
+      "id": "parallax_new_dawn",
+      "name": "New Dawn Hall Parallax",
+      "filename": "parallax/new_dawn_hall_parallax_background.png",
+      "status": "complete",
+      "config": {
+        "aspectRatio": "16:9",
+        "outputMimeType": "image/png"
+      },
+      "prompt": {
+        "scene": "Great Hall at sunrise after victory",
+        "setting": "Great Hall restored, Everember burning bright, dawn light streaming through tall windows",
+        "composition": "Pillars foreground, Everember midground, windows with sunrise background",
+        "lighting": "Golden sunrise mixing with warm firelight, hopeful and radiant",
+        "style": "Storybook painterly, triumphant and peaceful"
+      }
+    }
+  ]
+}
