+++ game/src/game/factories/enemy-factory.ts
@@ -11,7 +11,7 @@ import type { Enemy, EnemyType } from '../types/entities';
 import type { EnemyVehicle } from '../types/ai';
 import { getMatterModules } from '../physics/matter-wrapper';
 import { Vector3 } from 'yuka';
-import { getChapterManifestSync, getEnemiesManifestSync } from '../../../ddl/loader';
+import { getChapterManifestSync, getEnemiesManifestSync } from '@ddl/loader';
 
 /**
  * Enemy definition from chapter manifest
