+++ game/src/game/factories/npc-factory.ts
@@ -7,7 +7,7 @@
  */
 
 import { Vector3 } from 'yuka';
-import { getChapterManifestSync, getNPCsManifestSync } from '../../../ddl/loader';
+import { getChapterManifestSync, getNPCsManifestSync } from '@ddl/loader';
 
 /**
  * Build NPCs for a chapter
