+++ game/tsconfig.json
@@ -6,6 +6,7 @@
     "paths": {
       "@/*": ["./*"],
       "@game/*": ["./game/*"],
+      "@ddl/*": ["./ddl/*"],
       "@data/*": ["./data/*"]
     }
   },
